Moved year specific auto commands
