+++ src/team2202/robot/autoCommands/ContinuousPegVisionCommand.java
@@ -1,8 +1,9 @@
-package auto.commands;
+package team2202.robot.autoCommands;
 
 import java.util.ArrayList;
 
 import auto.ICommand;
+import auto.commands.DriveAtAngle;
 import auto.stopConditions.TimerStopCondition;
 import comms.NetworkTables;
 import comms.SmartWriter;
